Fixed multiplayer for localhost: the actual ip is used, moved get_ip to run.py, changed all environ.get failure argument to None. Hardcoded sizes for map to make it work for the moment.
