+++ run.py
@@ -6,7 +6,23 @@ import subprocess
 import sys
 import time
 import traceback
+import socket
+
 from subprocess import CalledProcessError
+
+def get_ip():
+    # http://stackoverflow.com/a/28950776/671626
+    s = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
+    try:
+        # doesn't even have to be reachable
+        s.connect(('10.255.255.255', 0))
+        IP = s.getsockname()[0]
+    except:
+        IP = '127.0.0.1'
+    finally:
+        s.close()
+    return IP
+
 
 _SCRIPT_LOCATION = os.path.abspath(os.path.dirname(__file__))
 _MANAGE_PY = os.path.join(_SCRIPT_LOCATION, 'example_project', 'manage.py')
@@ -71,8 +87,10 @@ def main(use_minikube):
         os.environ['AIMMO_MODE'] = 'minikube'
     else:
         time.sleep(2)
-        game = run_command_async(['python', _SERVICE_PY, '127.0.0.1', '5000'])
         os.environ['AIMMO_MODE'] = 'threads'
+        os.environ['GAME_API_URL'] = 'http://' + get_ip() + ":8000/players/api/games/"
+        os.environ['WORKER_MANAGER'] = 'local'
+        game = run_command_async(['python', _SERVICE_PY, get_ip(), '5000'])
         server_args.append('0.0.0.0:8000')
     server = run_command_async(['python', _MANAGE_PY, 'runserver'] + server_args)
